mission updat

+++ MossadApi/Models/Mission.cs
@@ -8,8 +8,8 @@ namespace MossadApi.Models
         public int Id { get; set; }
         public int AgentId { get; set; }
         public int TargetId { get; set; }
-        public float Timelaft { get; set; }
-        public float TotalTime { get; set; }
+        public double Timelaft { get; set; }
+        public double TotalTime { get; set; }
 
         [AllowedValues("assigned", "finish", "pussible")]
         public string Status { get; set; }
